cleanup old version files; experiment with new <Sprite />
