+++ packages/react-gamin/index.tsx
@@ -1,5 +1,3 @@
-export * from "./src/core/audio";
-export * from "./src/core/entities";
-export * from "./src/core/game";
-export * from "./src/core/input";
-export * from "./src/core/render";
+export * from "./src/audio";
+export * from "./src/input";
+export * from "./src/render";
